tweak imports to fix update contact config
